# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prediction Markets API",
    "description": "REST API for prediction market data and user features",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/markets": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "List Markets",
        "description": "List markets with optional filtering and pagination.\n\nReturns a paginated list of markets with their outcomes. Markets are ordered\nby last update time (most recent first).\n\n- **category**: Filter by event category (exact match)\n- **search**: Search text in market question and description (case-insensitive)\n- **resolved**: Filter by resolved status (true/false)\n- **limit**: Maximum number of markets to return (1-500, default 50)\n- **offset**: Number of markets to skip for pagination",
        "operationId": "list_markets_api_v1_markets_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (exact match)",
              "title": "Category"
            },
            "description": "Filter by category (exact match)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in question and description",
              "title": "Search"
            },
            "description": "Search in question and description"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolved status",
              "title": "Resolved"
            },
            "description": "Filter by resolved status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum results to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/markets/advanced": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "List Markets Advanced",
        "description": "List markets with real-time CLOB bid/ask data.\n\nReturns a paginated list of markets enriched with order book data from\nthe Polymarket CLOB API. This endpoint is suitable for advanced filtering\nand screening use cases.\n\n**Sort options:**\n- `volume`: Sort by 24-hour trading volume (default)\n- `liquidity`: Sort by estimated liquidity\n- `spread`: Sort by bid-ask spread\n- `priceChange`: Sort by 24-hour price change magnitude\n\n**Advanced filters:**\n- Volume range filters (minVolume, maxVolume)\n- Liquidity range filters (minLiquidity, maxLiquidity)\n- Spread range filters in cents (minSpread, maxSpread)\n- Price change filters (minChange, maxChange)\n\n**CLOB data fields:**\n- `bid`: Best bid price (0-1)\n- `ask`: Best ask price (0-1)\n- `spread`: Absolute spread (ask - bid)\n- `spreadPercent`: Spread as percentage of midpoint\n- `spreadCents`: Spread in cents\n- `volume24h`: 24-hour trading volume in USD\n- `priceChange24h`: 24-hour price change\n\nNote: CLOB data is cached for 30 seconds to reduce API calls.",
        "operationId": "list_markets_advanced_api_v1_markets_advanced_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (exact match)",
              "title": "Category"
            },
            "description": "Filter by category (exact match)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in question and description",
              "title": "Search"
            },
            "description": "Search in question and description"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolved status",
              "title": "Resolved"
            },
            "description": "Filter by resolved status"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "volume",
                "liquidity",
                "spread",
                "priceChange"
              ],
              "type": "string",
              "description": "Sort field",
              "default": "volume",
              "title": "Sortby"
            },
            "description": "Sort field"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sortorder"
            },
            "description": "Sort direction"
          },
          {
            "name": "minVolume",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 24h volume in USD",
              "title": "Minvolume"
            },
            "description": "Minimum 24h volume in USD"
          },
          {
            "name": "maxVolume",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum 24h volume in USD",
              "title": "Maxvolume"
            },
            "description": "Maximum 24h volume in USD"
          },
          {
            "name": "minLiquidity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum liquidity",
              "title": "Minliquidity"
            },
            "description": "Minimum liquidity"
          },
          {
            "name": "maxLiquidity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum liquidity",
              "title": "Maxliquidity"
            },
            "description": "Maximum liquidity"
          },
          {
            "name": "minSpread",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum spread in cents",
              "title": "Minspread"
            },
            "description": "Minimum spread in cents"
          },
          {
            "name": "maxSpread",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum spread in cents",
              "title": "Maxspread"
            },
            "description": "Maximum spread in cents"
          },
          {
            "name": "minChange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 24h price change",
              "title": "Minchange"
            },
            "description": "Minimum 24h price change"
          },
          {
            "name": "maxChange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum 24h price change",
              "title": "Maxchange"
            },
            "description": "Maximum 24h price change"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated tags to filter by",
              "title": "Tags"
            },
            "description": "Comma-separated tags to filter by"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum results to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of results to skip"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page number (alternative to offset)",
              "title": "Page"
            },
            "description": "Page number (alternative to offset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketAdvancedListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/markets/{market_id}": {
      "get": {
        "tags": [
          "markets"
        ],
        "summary": "Get Market",
        "description": "Get a single market by ID.\n\nThe market_id can be any of:\n- Internal market ID (platform_market_id)\n- Polymarket ID (same as platform_market_id for Polymarket markets)\n- Market slug (URL-friendly identifier)\n\nReturns the market with all its outcome details.",
        "operationId": "get_market_api_v1_markets__market_id__get",
        "parameters": [
          {
            "name": "market_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Market identifier (can be internal ID, polymarketId, or slug)",
              "title": "Market Id"
            },
            "description": "Market identifier (can be internal ID, polymarketId, or slug)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketAdvancedListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MarketAdvancedResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "MarketAdvancedListResponse",
        "description": "Paginated advanced market list response."
      },
...
      "MarketListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MarketResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "MarketListResponse",
        "description": "Paginated market list response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}